requirements.txt
google-adk~=1.2.1
black~=25.1.0
streamlit~=1.45.1
mysql-connector-python~=9.3.0
bcrypt~=4.3.0
watchdog~=6.0.0
